+++ changelog.md
@@ -1,3 +1,6 @@
+# 1.0.2
+- Add setting to add additional boilerplate triggers
+	- These triggers can make using these spawn loops easier in platformer levels
 # 1.0.1
 - Add automatic delay calculation
 	- It will automatically set the delay to go to the last selected object